daemon/session2: put event handling into separate functions
Originally, events were simply just another payload for the regular data
callbacks. That could, however, lead to problems with the control flow
of the layer iterations, e.g. when a buffer causes a layer to `_WAIT`,
but the next payload is an event instead of another buffer. This
separation should fix that problem.